# openapi_spec.yaml
openapi: 3.0.3
info:
  title: CliniSync / ShebaSync API Contract
  description: Official locked REST API specification for Frontend (Next.js & Flutter)
    and Backend (FastAPI).
  version: 1.0.0
servers:
- url: https://api.clinisync.com/api/v1
  description: Production Server
- url: http://localhost:8000/api/v1
  description: Local Development Server
paths:
...
  /availability/{doctorId}:
    get:
      summary: Compute free slots for a doctor on a specific date
      tags:
      - Availability
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: date
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        '200':
          description: Computed slots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilitySlotsResponse'
  /appointments/book:
    post:
      summary: Book an appointment (Idempotent)
      tags:
      - Appointments
      security:
      - BearerAuth: []
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookAppointmentRequest'
      responses:
        '201':
          description: Appointment booked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '409':
          description: Slot conflict (Already booked or overlapping)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /appointments/{id}/reschedule:
    put:
      summary: Reschedule appointment atomically
      tags:
      - Appointments
      security:
      - BearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - newStartTime
              - newEndTime
              properties:
                newStartTime:
                  type: string
                  format: date-time
                newEndTime:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Rescheduled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /appointments/{id}/cancel:
    put:
      summary: Cancel appointment
      tags:
      - Appointments
      security:
      - BearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Cancelled successfully
  /appointments/{id}/status:
    put:
      summary: Doctor toggles completed or no_show
      tags:
      - Appointments
      security:
      - BearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - status
              properties:
                status:
                  type: string
                  enum:
                  - completed
                  - no_show
      responses:
        '200':
          description: Status updated
  /appointments/doctor/{doctorId}:
    get:
      summary: Doctor daily queue
      tags:
      - Appointments
      security:
      - BearerAuth: []
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: date
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        '200':
          description: Daily queue list
  /appointments/patient/me:
    get:
      summary: Patient appointment list
      tags:
      - Appointments
      security:
      - BearerAuth: []
      parameters:
      - name: filter
        in: query
        schema:
          type: string
          enum:
          - upcoming
          - past
      responses:
        '200':
          description: Patient bookings
  /admin/analytics:
    get:
      summary: Get clinic volume and no-show metrics
      tags:
      - Admin
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Analytics data
  /admin/audit-log:
    get:
      summary: View paginated audit logs
      tags:
      - Admin
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Audit logs
...
  /availability/{doctorId}/recurring:
    put:
      summary: Set doctor weekly recurring availability
      tags:
      - Availability
      security:
      - BearerAuth: []
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - shifts
              properties:
                shifts:
                  type: array
                  items:
                    type: object
                    properties:
                      dayOfWeek:
                        type: integer
                        minimum: 0
                        maximum: 6
                      startTime:
                        type: string
                        example: 09:00
                      endTime:
                        type: string
                        example: '13:00'
                      slotDurationMinutes:
                        type: integer
                        default: 30
                      bufferMinutes:
                        type: integer
                        default: 10
      responses:
        '200':
          description: Weekly shifts updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /availability/{doctorId}/exceptions:
    post:
      summary: Block a specific date for holiday/leave
      tags:
      - Availability
      security:
      - BearerAuth: []
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - exceptionDate
              - isAvailable
              properties:
                exceptionDate:
                  type: string
                  format: date
                isAvailable:
                  type: boolean
                  default: false
                reason:
                  type: string
      responses:
        '201':
          description: Exception blocked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  doctorId:
                    type: string
                    format: uuid
                  exceptionDate:
                    type: string
                    format: date
                  isAvailable:
                    type: boolean
                  reason:
                    type: string
  /availability/{doctorId}/bulk-generate:
    post:
      summary: Bulk publish slots for next week or next month
      tags:
      - Availability
      security:
      - BearerAuth: []
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - targetRange
              - startDate
              - endDate
              properties:
                targetRange:
                  type: string
                  enum:
                  - week
                  - month
                  - custom
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
                daysOfWeek:
                  type: array
                  items:
                    type: integer
                startTime:
                  type: string
                  example: 09:00
                endTime:
                  type: string
                  example: '13:00'
                slotDurationMinutes:
                  type: integer
                  default: 30
                bufferMinutes:
                  type: integer
                  default: 10
                applyHolidays:
                  type: boolean
                  default: true
      responses:
        '201':
          description: Slots bulk generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalSlotsCreated:
                    type: integer
                  dateRange:
                    type: string
                  workingDaysCount:
                    type: integer
                  holidaysSkipped:
                    type: integer
                  slotsPerDay:
                    type: integer
                  message:
                    type: string
  /appointments/doctor/{doctorId}/queue-pause:
    put:
      summary: Hold queue for doctor break (enforces empty chamber invariant)
      tags:
      - Appointments
      security:
      - BearerAuth: []
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - pauseMinutes
              properties:
                pauseMinutes:
                  type: integer
                  default: 5
                reason:
                  type: string
      responses:
        '200':
          description: Chamber paused and previous patient completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  isPaused:
                    type: boolean
                  chamberStatus:
                    type: string
                    example: EMPTY (BREAK)
                  pauseMinutes:
                    type: integer
                  resumeAt:
                    type: string
                    format: date-time
                  message:
                    type: string
  /appointments/doctor/{doctorId}/queue-resume:
    put:
      summary: Resume chamber queue and call next patient token
      tags:
      - Appointments
      security:
      - BearerAuth: []
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Chamber resumed and next serial called
          content:
            application/json:
              schema:
                type: object
                properties:
                  isPaused:
                    type: boolean
                  currentlyServingSerial:
                    type: integer
                  chamberStatus:
                    type: string
                    example: ACTIVE
                  message:
                    type: string
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ErrorResponse:
      description: Error occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    AvailabilitySlotsResponse:
      type: object
      properties:
        doctorId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        slotDurationMinutes:
          type: integer
        bufferMinutes:
          type: integer
        slots:
          type: array
          items:
            type: object
            properties:
              startTime:
                type: string
                format: date-time
              endTime:
                type: string
                format: date-time
              isAvailable:
                type: boolean
    BookAppointmentRequest:
      type: object
      required:
      - doctorId
      - startTime
      - endTime
      properties:
        doctorId:
          type: string
          format: uuid
        startTime:
          type: string
          format: date-time
          example: '2026-09-10T10:00:00Z'
        endTime:
          type: string
          format: date-time
          example: '2026-09-10T10:30:00Z'
    Appointment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        doctorId:
          type: string
          format: uuid
        doctorName:
          type: string
        specialization:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - pending
          - confirmed
          - completed
          - cancelled
          - no_show
        createdAt:
          type: string
          format: date-time
    ErrorEnvelope:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - status
          properties:
            code:
              type: string
              example: SLOT_CONFLICT
            message:
              type: string
              example: This slot was just booked by another patient.
            status:
              type: integer
              example: 409
            details:
              type: object
...